+++ clients/src/main/test/com/amigoscode/clients/fraud/FraudCheckResponseTest.java
@@ -1,0 +1,34 @@
+package com.amigoscode.clients.fraud;
+
+import org.junit.jupiter.api.*;
+
+class FraudCheckResponseTest {
+
+    @BeforeEach
+    void setUp() {
+
+    }
+
+    @AfterEach
+    void tearDown() {
+
+    }
+
+    @BeforeAll
+    static void beforeAll() {
+
+    }
+
+    @AfterAll
+    static void afterAll() {
+
+    }
+
+    @Test
+    void name() {
+    }
+
+    @Test
+    void name2() {
+    }
+}
