+++ customer/src/main/java/com/amigoscode/customer/Customer.java
@@ -31,4 +31,5 @@ public class Customer {
     private String firstName;
     private String lastName;
     private String email;
+
 }
