+++ customer/src/main/java/com/amigoscode/customer/CustomerApplication.java
@@ -23,5 +23,6 @@ import org.springframework.context.annotation.PropertySources;
 public class CustomerApplication {
     public static void main(String[] args) {
         SpringApplication.run(CustomerApplication.class, args);
+
     }
 }
