+++ customer/src/main/java/com/amigoscode/customer/CustomerService.java
@@ -1,49 +1,5 @@
 package com.amigoscode.customer;
 
-import com.amigoscode.amqp.RabbitMQMessageProducer;
-import com.amigoscode.clients.fraud.FraudCheckResponse;
-import com.amigoscode.clients.fraud.FraudClient;
-import com.amigoscode.clients.notification.NotificationClient;
-import com.amigoscode.clients.notification.NotificationRequest;
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@AllArgsConstructor
-public class CustomerService {
-
-    private final CustomerRepository customerRepository;
-    private final FraudClient fraudClient;
-    private final RabbitMQMessageProducer rabbitMQMessageProducer;
-
-    public void registerCustomer(CustomerRegistrationRequest request) {
-        Customer customer = Customer.builder()
-                .firstName(request.firstName())
-                .lastName(request.lastName())
-                .email(request.email())
-                .build();
-        // todo: check if email valid
-        // todo: check if email not taken
-        customerRepository.saveAndFlush(customer);
-
-        FraudCheckResponse fraudCheckResponse =
-                fraudClient.isFraudster(customer.getId());
-
-        if (fraudCheckResponse.isFraudster()) {
-            throw new IllegalStateException("fraudster");
-        }
-
-        NotificationRequest notificationRequest = new NotificationRequest(
-                customer.getId(),
-                customer.getEmail(),
-                String.format("Hi %s, welcome to Amigoscode...",
-                        customer.getFirstName())
-        );
-        rabbitMQMessageProducer.publish(
-                notificationRequest,
-                "internal.exchange",
-                "internal.notification.routing-key"
-        );
-
-    }
+public interface CustomerService {
+    void registerCustomer(CustomerRegistrationRequest request);
 }
