+++ customer/src/main/test/com/amigoscode/customer/CustomerTest.java
@@ -1,0 +1,16 @@
+package com.amigoscode.customer;
+
+import org.junit.jupiter.api.Test;
+
+class CustomerTest {
+
+    @Test
+    void name() {
+        //given
+
+        //when
+
+        //yes
+
+    }
+}
