+++ fraud/src/main/java/com/amigoscode/fraud/FraudCheckService.java
@@ -1,25 +1,5 @@
 package com.amigoscode.fraud;
 
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
-
-@Service
-@AllArgsConstructor
-public class FraudCheckService {
-
-    private final FraudCheckHistoryRepository fraudCheckHistoryRepository;
-
-    public boolean isFraudulentCustomer(Integer customerId) {
-        fraudCheckHistoryRepository.save(
-                FraudCheckHistory.builder()
-                        .customerId(customerId)
-                        .isFraudster(false)
-                        .createdAt(LocalDateTime.now())
-                        .build()
-        );
-        return false;
-    }
-
+public interface FraudCheckService {
+    boolean isFraudulentCustomer(Integer customerId);
 }
