+++ notification/src/main/resources/application-docker.yml
@@ -1,4 +1,4 @@
-server:
+clients:
     port: 8082
 spring:
     application:
@@ -22,7 +22,7 @@ spring:
 eureka:
     client:
         service-url:
-            defaultZone: http://eureka-server:8761/eureka
+            defaultZone: http://eureka-clients:8761/eureka
         fetch-registry: true
         register-with-eureka: true
         enabled: false
