+++ notification/src/main/resources/application.yml
@@ -1,4 +1,4 @@
-server:
+clients:
     port: 8082
 spring:
     application:
