+++ notification/src/main/test/com/amigoscode/notification/NotificationServiceTest.java
@@ -1,0 +1,58 @@
+package com.amigoscode.notification;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class NotificationServiceTest {
+
+    @BeforeEach
+    void setUp() {
+    }
+
+    @AfterEach
+    void tearDown() {
+    }
+
+    @Test
+    void send() {
+
+
+    }
+
+    @Test
+    void send2() {
+
+
+    }
+
+    @Test
+    void send3() {
+        //given
+        //when
+        //then
+
+    }
+
+    @Test
+    void send4() {
+        //given
+        //when
+        //then
+
+    }
+
+    @BeforeAll
+    static void beforeAll() {
+
+    }
+
+    @Test
+    void name() {
+        //given
+        //when
+        //then
+
+    }
+}
